revert Specs etc to original location
