+++ crates/flux-middle/src/lib.rs
@@ -40,19 +40,23 @@ pub mod timings;
 use std::sync::LazyLock;
 
 use flux_arc_interner::List;
+use flux_config as config;
 use flux_macros::fluent_messages;
 pub use flux_rustc_bridge::def_id_to_string;
 use flux_rustc_bridge::{
     mir::{LocalDecls, PlaceElem},
     ty::{self, GenericArgsExt},
 };
-use flux_syntax::surface::NodeId;
+use flux_syntax::surface::{self, NodeId};
 use global_env::GlobalEnv;
 use liquid_fixpoint::ThyFunc;
 use queries::QueryResult;
 use rty::VariantIdx;
 use rustc_abi::FieldIdx;
-use rustc_data_structures::unord::UnordMap;
+use rustc_data_structures::{
+    fx::FxIndexMap,
+    unord::{UnordMap, UnordSet},
+};
 use rustc_hir::OwnerId;
 use rustc_macros::extension;
 use rustc_middle::ty::TyCtxt;
@@ -312,6 +316,77 @@ fn sort_of_thy_func(func: liquid_fixpoint::ThyFunc) -> Option<rty::PolyFuncSort>
         }
     };
     Some(sort)
+}
+
+#[derive(Debug, Default)]
+pub struct Specs {
+    pub fn_sigs: UnordMap<OwnerId, surface::FnSpec>,
+    pub constants: UnordMap<OwnerId, surface::ConstantInfo>,
+    pub structs: UnordMap<OwnerId, surface::StructDef>,
+    pub traits: UnordMap<OwnerId, surface::Trait>,
+    pub impls: UnordMap<OwnerId, surface::Impl>,
+    pub enums: UnordMap<OwnerId, surface::EnumDef>,
+    pub flux_items_by_parent: FxIndexMap<OwnerId, Vec<surface::Item>>,
+    pub ty_aliases: UnordMap<OwnerId, Option<surface::TyAlias>>,
+    pub ignores: UnordMap<LocalDefId, fhir::Ignored>,
+    pub trusted: UnordMap<LocalDefId, fhir::Trusted>,
+    pub trusted_impl: UnordMap<LocalDefId, fhir::Trusted>,
+    pub infer_opts: UnordMap<LocalDefId, config::PartialInferOpts>,
+    pub should_fail: UnordSet<LocalDefId>,
+    /// Set of dummy items generated by the extern spec macro we must completely ignore. This is
+    /// not the same as [ignored items] because, for ignored items, we still need to return errors
+    /// for queries and handle them gracefully in order to report them at the use it.
+    ///
+    /// If an item is in this set, all its descendants are also consider dummy (but they may not be
+    /// in the set).
+    ///
+    /// [ignored items]: Specs::ignores
+    pub dummy_extern: UnordSet<LocalDefId>,
+    pub extern_id_to_local_id: UnordMap<DefId, LocalDefId>,
+    pub local_id_to_extern_id: UnordMap<LocalDefId, DefId>,
+}
+
+impl Specs {
+    pub fn insert_extern_spec_id_mapping(
+        &mut self,
+        local_id: LocalDefId,
+        extern_id: DefId,
+    ) -> Result<(), ExternSpecMappingErr> {
+        #[expect(
+            clippy::disallowed_methods,
+            reason = "we are inserting the extern spec mapping and we want to ensure it doesn't point to a local item"
+        )]
+        if let Some(local) = extern_id.as_local() {
+            return Err(ExternSpecMappingErr::IsLocal(local));
+        }
+        if let Err(err) = self.extern_id_to_local_id.try_insert(extern_id, local_id) {
+            return Err(ExternSpecMappingErr::Dup(*err.entry.get()));
+        }
+        self.local_id_to_extern_id.insert(local_id, extern_id);
+        Ok(())
+    }
+
+    pub fn insert_dummy(&mut self, owner_id: OwnerId) {
+        self.dummy_extern.insert(owner_id.def_id);
+    }
+}
+
+/// Represents errors that can occur when inserting a mapping between a `LocalDefId` and a `DefId`
+/// for an extern spec.
+pub enum ExternSpecMappingErr {
+    /// Indicates that the [`DefId`] we are trying to add extern specs to is actually local. Returns
+    /// the [`DefId`] as a [`LocalDefId`].
+    IsLocal(LocalDefId),
+
+    /// Indicates that there is an existing extern spec for the given extern id. Returns the existing
+    /// `LocalDefId` that maps to the extern id.
+    ///
+    /// NOTE: This currently only considers extern specs defined in the local crate. There could still
+    /// be duplicates if an extern spec is imported from an external crate. In such cases, the local
+    /// extern spec takes precedence. Probably, we should at least warn about this, but it's a bit
+    /// tricky because we need to look at the crate metadata which we don't have handy when
+    /// collecting specs.
+    Dup(LocalDefId),
 }
 
 #[derive(Default)]
