+++ crates/flux-syntax/src/surface.rs
@@ -1,16 +1,10 @@
 pub mod visit;
 use std::{borrow::Cow, fmt, ops::Range};
 
-use flux_config::PartialInferOpts;
 pub use rustc_ast::{
     Mutability,
     token::{Lit, LitKind},
 };
-use rustc_data_structures::{
-    fx::FxIndexMap,
-    unord::{UnordMap, UnordSet},
-};
-use rustc_hir::OwnerId;
 pub use rustc_span::{
     Span,
     def_id::{DefId, LocalDefId},
@@ -28,120 +22,6 @@ pub struct NodeId(pub(super) usize);
 impl NodeId {
     pub fn as_usize(&self) -> usize {
         self.0
-    }
-}
-
-/// Represents errors that can occur when inserting a mapping between a `LocalDefId` and a `DefId`
-/// for an extern spec.
-pub enum ExternSpecMappingErr {
-    /// Indicates that the [`DefId`] we are trying to add extern specs to is actually local. Returns
-    /// the [`DefId`] as a [`LocalDefId`].
-    IsLocal(LocalDefId),
-
-    /// Indicates that there is an existing extern spec for the given extern id. Returns the existing
-    /// `LocalDefId` that maps to the extern id.
-    ///
-    /// NOTE: This currently only considers extern specs defined in the local crate. There could still
-    /// be duplicates if an extern spec is imported from an external crate. In such cases, the local
-    /// extern spec takes precedence. Probably, we should at least warn about this, but it's a bit
-    /// tricky because we need to look at the crate metadata which we don't have handy when
-    /// collecting specs.
-    Dup(LocalDefId),
-}
-
-/// A boolean-like enum used to mark whether a piece of code is ignored.
-#[derive(Debug, Eq, PartialEq, Copy, Clone)]
-pub enum Ignored {
-    Yes,
-    No,
-}
-
-impl Ignored {
-    pub fn to_bool(self) -> bool {
-        match self {
-            Ignored::Yes => true,
-            Ignored::No => false,
-        }
-    }
-}
-
-impl From<bool> for Ignored {
-    fn from(value: bool) -> Self {
-        if value { Ignored::Yes } else { Ignored::No }
-    }
-}
-
-/// A boolean-like enum used to mark whether some code should be trusted.
-#[derive(Debug, Eq, PartialEq, Copy, Clone)]
-pub enum Trusted {
-    Yes,
-    No,
-}
-
-impl Trusted {
-    pub fn to_bool(self) -> bool {
-        match self {
-            Trusted::Yes => true,
-            Trusted::No => false,
-        }
-    }
-}
-
-impl From<bool> for Trusted {
-    fn from(value: bool) -> Self {
-        if value { Trusted::Yes } else { Trusted::No }
-    }
-}
-#[derive(Debug, Default)]
-pub struct Specs {
-    pub fn_sigs: UnordMap<OwnerId, FnSpec>,
-    pub constants: UnordMap<OwnerId, ConstantInfo>,
-    pub structs: UnordMap<OwnerId, StructDef>,
-    pub traits: UnordMap<OwnerId, Trait>,
-    pub impls: UnordMap<OwnerId, Impl>,
-    pub enums: UnordMap<OwnerId, EnumDef>,
-    pub flux_items_by_parent: FxIndexMap<OwnerId, Vec<Item>>,
-    pub ty_aliases: UnordMap<OwnerId, Option<TyAlias>>,
-    pub ignores: UnordMap<LocalDefId, Ignored>,
-    pub trusted: UnordMap<LocalDefId, Trusted>,
-    pub trusted_impl: UnordMap<LocalDefId, Trusted>,
-    pub infer_opts: UnordMap<LocalDefId, PartialInferOpts>,
-    pub should_fail: UnordSet<LocalDefId>,
-    /// Set of dummy items generated by the extern spec macro we must completely ignore. This is
-    /// not the same as [ignored items] because, for ignored items, we still need to return errors
-    /// for queries and handle them gracefully in order to report them at the use it.
-    ///
-    /// If an item is in this set, all its descendants are also consider dummy (but they may not be
-    /// in the set).
-    ///
-    /// [ignored items]: Specs::ignores
-    pub dummy_extern: UnordSet<LocalDefId>,
-    pub extern_id_to_local_id: UnordMap<DefId, LocalDefId>,
-    pub local_id_to_extern_id: UnordMap<LocalDefId, DefId>,
-}
-
-impl Specs {
-    pub fn insert_extern_spec_id_mapping(
-        &mut self,
-        local_id: LocalDefId,
-        extern_id: DefId,
-    ) -> Result<(), ExternSpecMappingErr> {
-        #[expect(
-            clippy::disallowed_methods,
-            reason = "we are inserting the extern spec mapping and we want to ensure it doesn't point to a local item"
-        )]
-        if let Some(local) = extern_id.as_local() {
-            return Err(ExternSpecMappingErr::IsLocal(local));
-        }
-        if let Err(err) = self.extern_id_to_local_id.try_insert(extern_id, local_id) {
-            return Err(ExternSpecMappingErr::Dup(*err.entry.get()));
-        }
-        self.local_id_to_extern_id.insert(local_id, extern_id);
-        Ok(())
-    }
-
-    pub fn insert_dummy(&mut self, owner_id: OwnerId) {
-        self.dummy_extern.insert(owner_id.def_id);
     }
 }
